Move small hemi logo in token to its own component
